data-generator-module/requirement.txt
azure-iot-device~=2.5.0
numpy